Refactor code after recommendations
