+++ healthtools/tests/test_scrapers.py
@@ -5,9 +5,9 @@ from healthtools.scrapers.doctors import DoctorsScraper
 from healthtools.scrapers.foreign_doctors import ForeignDoctorsScraper
 from healthtools.scrapers.clinical_officers import ClinicalOfficersScraper
 from healthtools.scrapers.health_facilities import HealthFacilitiesScraper
-from healthtools.scrapers.nhif_outpatient_cs_facilities import NhifOutpatientCsScraper
-from healthtools.scrapers.nhif_inpatient_facilities import NhifInpatientScraper
-from healthtools.scrapers.nhif_outpatient_facilities import NhifOutpatientScraper
+from healthtools.scrapers.nhif_inpatient import NhifInpatientScraper
+from healthtools.scrapers.nhif_outpatient import NhifOutpatientScraper
+from healthtools.scrapers.nhif_outpatient_cs import NhifOutpatientCsScraper
 from healthtools.config import TEST_DIR, SLACK, AWS
 
 
@@ -18,7 +18,7 @@ class TestScrapers(unittest.TestCase):
         self.foreign_doctors_scraper = ForeignDoctorsScraper()
         self.clinical_officers_scraper = ClinicalOfficersScraper()
         self.health_facilities_scraper = HealthFacilitiesScraper()
-        self.nhif_accredited_scraper = NhifOutpatientCsScraper()
+        self.nhif_outpatient_cs_scraper = NhifOutpatientCsScraper()
         self.nhif_inpatient_scraper = NhifInpatientScraper()
         self.nhif_outpatient_scraper = NhifOutpatientScraper()
 
@@ -42,7 +42,7 @@ class TestScrapers(unittest.TestCase):
         self.assertTrue(len(entries) == 60)
 
     def test_it_scrapes_nhif_accredited_page(self):
-        entries = self.nhif_accredited_scraper.scrape_page(1)
+        entries = self.nhif_outpatient_cs_scraper.scrape_page(1)
         self.assertTrue(len(entries) > 1)
 
     def test_it_scrapes_nhif_inpatient_page(self):
@@ -78,19 +78,19 @@ class TestScrapers(unittest.TestCase):
             self.assertEqual(response["items"][0]["index"]["_shards"]["successful"], 1)
 
     def test_nhif_accredited_scraper_uploads_to_elasticsearch(self):
-        with open(TEST_DIR + "/dummy_files/nhif_accredited_facilities.json", "r") as my_file:
+        with open(TEST_DIR + "/dummy_files/nhif_outpatient_cs.json", "r") as my_file:
             data = my_file.read()
-            response = self.nhif_accredited_scraper.upload_data(json.loads(data))
+            response = self.nhif_outpatient_cs_scraper.upload_data(json.loads(data))
             self.assertEqual(response["items"][0]["index"]["_shards"]["successful"], 1)
 
     def test_nhif_inpatient_scraper_uploads_to_elasticsearch(self):
-        with open(TEST_DIR + "/dummy_files/nhif_inpatient_facilities.json", "r") as my_file:
+        with open(TEST_DIR + "/dummy_files/nhif_inpatient.json", "r") as my_file:
             data = my_file.read()
             response = self.nhif_inpatient_scraper.upload_data(json.loads(data))
             self.assertEqual(response["items"][0]["index"]["_shards"]["successful"], 1)
 
     def test_nhif_outpatient_scraper_uploads_to_elasticsearch(self):
-        with open(TEST_DIR + "/dummy_files/nhif_outpatient_facilities.json", "r") as my_file:
+        with open(TEST_DIR + "/dummy_files/nhif_outpatient.json", "r") as my_file:
             data = my_file.read()
             response = self.nhif_outpatient_scraper.upload_data(json.loads(data))
             self.assertEqual(response["items"][0]["index"]["_shards"]["successful"], 1)
